fixed issues

+++ src/include/vec2.h
@@ -4,19 +4,19 @@
 class Vec2 {
     public:
         Vec2() = default;
-        Vec2(double _x, double _y);
+        Vec2(float _x, float _y);
 
         Vector2 toRaylib();
 
         Vec2& operator+=(const Vec2& v);
         Vec2& operator-=(const Vec2& v);
-        Vec2& operator*=(const double& k);
+        Vec2& operator*=(const float& k);
 
         ~Vec2() = default;
-        double x = 0, y = 0;
+        float x = 0, y = 0;
 };
 
 Vec2 operator+(const Vec2& v1, const Vec2& v2);
 Vec2 operator-(const Vec2& v1, const Vec2& v2);
-Vec2 operator*(const Vec2& v, const double& k);
-Vec2 operator*(const double& k, const Vec2& v);
+Vec2 operator*(const Vec2& v, const float& k);
+Vec2 operator*(const float& k, const Vec2& v);
